factor out the DbcSpecifics class from dbc.py
the motivation is to be able to import the DBC specifics class in
modules which also import dbc.py. (This is required for adding type
hints to the internal database code properly.)
